Changed the getValue of the ValueProvider because it always returned an object of type Any causing the failure of the check.
Now the getValue tries to recognize the type of the property and returns an object of the right type to the caller. Booleans are translated to true and false.
The following type are recognized at run time trying to convert the type object to one of them and returning the first that secceds:
Integer, Long, Float, Double, Boolean, String

I know that this is only a part of ll the types that a property can be but until now I have seen only String, boolean and integers in the properties generated at run time. However we should keep it in mind in case of problems with filtering.
Because of the SQL espressions and the properties of the examples I have, the method has been succesfully tested only against boolean and not existent properties.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@50206 523d945c-050c-4681-91ec-863ad3bb968a