Script to start netbeans with the GUI for the alarm system. Not yet completed because does not install the last version of the alarm gui in the modules dir of netbeans.
The script launch the GUI for the given ACS_INSTANCE or with a remote manager.
It takes th jars files from the right dirs (ACS policy) while building the classpath.
It assumes to found netbeans in  NBPLATFORM_HOME or in the home directory.

It starts the GUI but is not yet complete.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@50308 523d945c-050c-4681-91ec-863ad3bb968a