LogLevel defaults to INFO.
There is still something wrong in the setLogLevel because it seems to have a ghost filter that create problems while changing the log level from a bigger to a lower index (the user need to select twice the log level to make it works).
I have fixed this problem clearing the system filtres before changing the log level and it now works well. It can create problems in feature if we add more SystemFilters.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@50502 523d945c-050c-4681-91ec-863ad3bb968a