The LogEntry now exists in 2 versions: one for XML and one without the Node.
This last version will replace the previous one to avoid to parse several time the same log entry.

A new version of the cache will also be implemented with a string and separators instead of XML.

An interface has been written to easily swap between one implementation and the other (in this version is not yet used)

All the files are updated fot the new files/packages.

The tests pass.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@51208 523d945c-050c-4681-91ec-863ad3bb968a