ACSRemoteLogListener expanded with callback for the status of the connection (connecting/connected/disconnected).
LCEngine owns a vector of listeners in order to execute the callback for all the registered listeners (this is done with a specific method for each possible callback)
Each callback can be executed by different objects/threads so they are exceuted synchronously.

LCEngine is passed to ACSRemoteAccess and ACSStructuredPushConsumer to perform callbacks.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@53379 523d945c-050c-4681-91ec-863ad3bb968a