Used a synchronized queue for the xmlLogs (safer and more readable source).
The queue of logs is limited in size. If jlog is not able to follow the flow of the rate arriving from the NC, the queue get empty and some log are discarded.
In this case a message is displayed in the stdout.

TODO: change the interface to let the listener acknowledge this abnormal situation.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@54362 523d945c-050c-4681-91ec-863ad3bb968a