The signarute of get_component/service is changed and it now returns exceptions in case of failure instead of a nil object. The code has been accordingly changed to catch the exceptions.
In the module it seems that there is no proper handling for ACS
exceptions.
When an exception is catched, simply a new RemoteException is created
and thrown, while we should instead or throw an ACS exception or log
it before throwisng something else.
In this way we lose the original error trace.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@60146 523d945c-050c-4681-91ec-863ad3bb968a