The layout of the page is ok now.
I had to move the JComboBoxes to the bottom of the dialog otherwise they do not popup.
I have found on the net that sometimes the combo boxes and the textfields in the same dialog can have a strange behaviour. They suggest to move or use a Frame instead of a Jpanel as a container for the combo boxes.
I hope the look and feel is good enough (I prefer to have the min and mox log type where they were before...)


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@61020 523d945c-050c-4681-91ec-863ad3bb968a