Buffer the logs while the interface is pasued. It is the same operation that is done while loading/saving logs and new logs are inserted.
The buffer is flushed whenever a long operation is terminated. To force the flush of the logs I have added a new LogOperationRequest called FLUSH_QUEUE to flush the queue asynchronously to avoid to block everything.
This operation is not esplicitly called in the thread (there is a comment in the code to explain the details).


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@63669 523d945c-050c-4681-91ec-863ad3bb968a