Added start/stop and pause/resume method to use as EXEC plugins
The empty constructor is now public because it is needed by the plugin.
It tries to check if another object of the same type has already been created (and stored as singleton).
This should be done in mutual exclusion to be 100% sure that it works.
I know there is something to address in future here.

See also alma.exec.extension.subsystemplugin.*


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@65318 523d945c-050c-4681-91ec-863ad3bb968a