When run by the acsdaemon, the Container would crash on startup because it was receiving an IOR instead of a corbaloc: string, which was used by the ACSCorba module to access ACS services directly.  After changing ACSCorba to direct its requests to the Manager, the Container was still crashing because no connection to the Manager was available.  Changing the order of initialization so that the connection to the Manager was made before any ACS service requests were attempted removed the failure.
If the Python component's CDB code information was incorrect, the
activate_component method would log the Python exception that caused
the fault.  This message gives no information to the user about what
the actual problem was without reading the Container source code.
An exception handler was added for the two Python exceptions that
are raised due to bad CDB information that raises ACSErrors with
more informative error messages.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@65355 523d945c-050c-4681-91ec-863ad3bb968a