COMP-1400, Intermittent problem connecting to the AMBSocketServer (Python component), appears to be caused by a failure to obtain the CORBA stubs for the requested object.  There are two causes for the failure that are unchecked: 1) the requested object isn't in the list of available components or 2) the CORBA stub module failed to load.
__importComponentStubs has been modified to throw a CannotGetComponentEx
if either error condition appears.  The documentation for the users
of this method have been modified to reflect the new behaviour.  Also
added are a set of unit tests that verify that the correct exception
is thrown in both cases.

Tests for None values have been changed to improve readability.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@76173 523d945c-050c-4681-91ec-863ad3bb968a