Python 2.5 has a couple of behavioural changes that affect the outputs and subsequently the test results.  For script that run in the mainline, the name is reported as <module> rather than ?.  Also, object references will report the values of the attributes instead of just the memory address.
The Name Service lookup routine was reporting a failure when it failed
to find a reference.  Instead of logging the error when not found, the
lookup routine simply passes the NotFound exception to the caller.

There are two persistent test failures that are addressed in these changes.
In one, the "Goodbye" message from the Container does not appear when it
should.  In the other, the order of the components returned has changed.
Flushing the stdout and sorting the returned list will hopefully eliminate
these minor errors.

Also, the CDB was updated to conform to the new Container schema.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@77049 523d945c-050c-4681-91ec-863ad3bb968a