C++ DevIOs return both a value and a timestamp, but Python DevIOs did not. (see COMP-1405 for more details)  This change takes a timestamp that was passed back from the DevIO in a tuple of (value, timestamp) and uses the timestamp given as the time in the Completion.  If the value is not a tuple, a timestamp will be generated for the Completion.
The BACI tests were enabled and modified to test the new behaviour.
The test IDL was updated so that a timestamp-returning property could
be accessed.

The NC channel test was removed from the Makefile.  The base script
is no longer part of the test suite, but it was still listed in the
scripts to be installed.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@77859 523d945c-050c-4681-91ec-863ad3bb968a