ACS 7.0 port.
Changelog:

   * Some GUI stylistic modifications, suggested by Erik Allaert
   * Change the imports (com.sun.* to org.apache.*) to avoid the warnings at
     compilation time
   * Fix the schemaLocation problem. Details:

Erik Allaert said:
''5. When trying to add the " Defaults " (according to the manual
" add all the errors files inside the ACS default errors files directories "),
I get a NullPointerException. Same if I try
to " Add " a directory - which means I didn't manage to test
the rest of the GUI ...''
Mauricio Araya response:
''The problem was that in ACS 6.0.4 (the ACS version of our development
VmWare machines), the XML files has the xsi:noNamespaceSchemaLocation
field, because known problem with the namespace. Nicolas Barriga told me that he built
a script to convert ACS 6.0.X error definitions to ACS 7.0 errors definitions,
because the current version includes xmlns and xsi:schemaLocation''


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@78378 523d945c-050c-4681-91ec-863ad3bb968a