Set the host name when processing an alarm.
The method updateAlarmHost set the name of the hot in the Source of the Alarm if it is not already set. In this way the name of the host is sent to the clients.
The name of the host is available in the message sent be sources to the alarm service (in the ASIMessage)


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@79348 523d945c-050c-4681-91ec-863ad3bb968a