COMP-1353 requests that Python containers implement dynamically configurable log levels.  These changes implement the LoggingConfigurable interface in the Container class as well as individually configurable loggers for components.
The loggers are organized in a tree structure based on the name.  A root logger,
called "DefaultPythonLogger", is created and holds the log level values from
the environment variables.  The container creates a default logger with the
container name as key.  It holds the log level values provided in the CDB.
Finally, component loggers are named "containername.componentname" and get
their log levels from specific values in the CDB or the parent nodes.

BaseClient would overwrite the logger from the Container class in its
constructor.  Now the BaseClient will only create a logger if one has
not been previously defined.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@80379 523d945c-050c-4681-91ec-863ad3bb968a