If a Container has been dynamically created and has no CDB information, the configureComponentLogger() method would fail because a CDB exception is received.  In that case, the correct behaviour is to set the component's log levels to be default.
ContainerServices objects were throwing exceptions because they had no
access to a logger object.  The logger object is now stored within
ContainerServices as well as being returned to the requesting component.

Duplicate log messages were appearing.  The Python logger hierarchy causes
a message to be logged at each level of the tree when the propagate attribute
is not 0.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@80519 523d945c-050c-4681-91ec-863ad3bb968a