Removed superfluous log message from Container's refresh_config method.
Many tests were failing because the component name was being reported
as Container.Component instead of the expected Component.  Since the
log record is used as the data source for all logging messages, the
Python LogRecord class was replaced with an ACSLogRecord which splits
the container name from the component name and stores the container
name so it can be loaded into the SourceObject field.  The makeRecord
method of the Logger class was overridden to return AcsLogRecords.

Updated reference files to reflect the current CDB settings.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@80586 523d945c-050c-4681-91ec-863ad3bb968a