The SmartPtr class stored the pointer to the name of the component that will be activated by getComponentSmartPtr().  If the name is deallocated, before the releaseComponent() call from within the SmartPtr, the operation will fail because the pointer is no longer valid.
Instead of storing the pointer, a deep copy of the component's name is
retained inside the ComponentStorage object.  This behaviour is consisted
across all of the component get operations.  Exception handling was also
added to prevent the failure of a component release from causing the
container to crash as well.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@81979 523d945c-050c-4681-91ec-863ad3bb968a