COMP-1296 (Creation of a dynamic python component property fails but component starts up ok) reported that a narrow operation failed when creating a property for a Python component.  It was determined that the Interface Repository was returning a NULL object and that an exception should be thrown to more clearly indicate a failure.
In the investigation of COMP-1774 (Investigate synchronization of
acsStart and Interface Repository loading), it was determined that
this operation was the only reference to the Interface Repository and
that the necessary information could be retrieved from the Python
module directly.  This change moves the information source from the
Interface Repository to the Python module.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@82244 523d945c-050c-4681-91ec-863ad3bb968a