COMP-1930 requests that exceptions be thrown if a user attempts to log at the OFF level.  logErrorTrace(), logTypeSafe() and logNotSoTypeSafe() were also changed to thrown exceptions if priority values that aren't defined in ACSLog.Priorities.
Erik also reported that the Python container is unable to log messages at
the ERROR level.  The code mapped any message received at the ERROR level
to the CRITICAL level because there was no method for logError() method
in the ACSLog.LogSvc interface.  Messages at the ERROR level are now
supported via a logError() method in the Logger class and transmitted via
a LogSvc logWithAudience() method.  The tat tests were updated to reflect
the new behaviour.

Finally, a unit test suite was implemented for the Acspy.Common.Log module
that does not depend on having ACS running.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@82920 523d945c-050c-4681-91ec-863ad3bb968a