To avoid connceting/disconnecting to the same NC each time a publish method with a Topic is called, the class declare a pool of publishers.
These methods are used by clients, categories and the heartbeat.
Sources instead use the constructor passing the topic to use. This channel is also added in the pool.

The class is ready to implement a mechanism to disconnect the publisher and free the resources for publisher not used for a long time.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@84884 523d945c-050c-4681-91ec-863ad3bb968a