COMP-1322 requests that all containers be configurable from the CDB. These changes implement the requested configurables in the Python container as module-level functions and a module-level thread to handle periodic flushing of the buffered messages.  In addition, the nesting of loggers has been replaced with an internal nesting of handlers for stdout and the central logging system.
Processing of error traces and type-safe log messages has been unified
with the regular log messages.  These messages, which were formerly
immediately sent to the log system, are now queued with the regular
messages.  The separate pending queue of error traces has been
eliminated along with the bug that assigns the priority of the
current error trace to the pending error traces.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@88852 523d945c-050c-4681-91ec-863ad3bb968a