The Python type-safe logging test in the loggingts module was failing to produce the expected output.  Within the test functions, the execution context and source information were being generated and passed into the Python logger where they were being silently dropped in favour of internally generated information.  The generated information did not match the output expected by the LogListener so the generated information was discarded.
This change passes context and source information through to the
handlers.  The formatter was updated to also accept data as a list
of NVPairs as well as a dictionary.  It was also noted that the
pending message buffer was not flushed when the interpreter exited
unless the periodic flush thread was operating.

Unit tests were updated as well.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@89585 523d945c-050c-4681-91ec-863ad3bb968a