The log type is displayed with the renderer in order to show the icon.
Each cell is displayed with a renderer that sets also the height of the table cell in order to show multi-line entries.
Note that the calculation of the height is not perfect and usually is a little bit greater then desired (I think it is ok as it is; I have found other way to do that on the net but all of them seem to work only up to jdk 1.3!).

The toottip is multiline with a max width of 96 chars.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@91466 523d945c-050c-4681-91ec-863ad3bb968a