COMP-2766 reported that the SmartPtrs were not released correctly when they go out of scope.  Upon further investigation, it was determined that because regular and SmartPtr operations were being mixed, SmartPtrs were being left with no ContainerServices reference that could be used to call releaseComponent().  As a result, the references were left dangling.
This change removes the ability to instantiate SmartPtrs from a component
reference and does not allow SmartPtrs with missing ContainerServices
references to be used in assignment or copy operations.  Hopefully,
this change will discourage users from mixing regular and SmartPtr
operations.

A test suite for the SmartPtr class was added as well.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@100903 523d945c-050c-4681-91ec-863ad3bb968a