COMP-2899 reports that the EpochHelper wrapper for Python reports an inconsistent number of digits for milliseconds, either dropping 1 from a 6 digit request or appending garbage characters to the string.  This behaviour caused the tests in the acspy module to fail on both 32-bit and 64-bit machines.
The problem was caused by interaction between an old wrapper specification
and the newer version of gcc using in ACS.  The old spec declared that the
toString methods of EpochHelper and DurationHelper would return char*, while
the underlying libraries were returning std::string objects.  The version of
SWIG shipped with ACS correctly flagged this error but, since the actual
wrapper had not been regenerated in over 2 years, this fault went unnoticed.
Correcting the return type in the spec allowed a correct wrapper to be
generated and eliminated the need to append a string terminator to the
returned value from the toString methods.

A side effect of the new wrapper was that the DurationHelper's positive
method started returning Boolean values rather than integers.  The reference
file in the acspy tests was updated to reflect the new behaviour.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@104133 523d945c-050c-4681-91ec-863ad3bb968a