COMP-3296 reports that acsstartupContainerPorts throws a ValueError exception when attempting to close the USED_CONTAINER_PORTS file. The close operation reported an IOError but apparently succeeded. At program exit, the cleanUp routine attempted to unlock a closed file causing the ValueError to be thrown.
The change improves the exception handling of the cleanUp routine.
If there is an exception, the operation is repeated after a timeout
period and a message written to stderr.  If, after the maximum timeout
period expires, the operation doesn't succeed, the program aborts.

Unit tests were updated to exercise the new code.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@108730 523d945c-050c-4681-91ec-863ad3bb968a