The parallelism has been revised by synchronizing at the level of the public methods. Some of the data structures have been revised accordingly because do not need anymore to be synchronized.
All the changes were needed to fix a bug (COMP-3528) but the problem was not really due to a missed synchronization but to the management of the file. In fact to reduce the amount of used disk space, the file was truncated when possible but the process still tried to read data from the old position.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@112072 523d945c-050c-4681-91ec-863ad3bb968a