CASA integration has been troublesome so we are taking a new approach.  Instead of simply adding CASA to the ACS environment used by everyone, we are making CASA access an option.  When selected, CASA directories are injected into the PYTHONPATH and LD_LIBRARY_PATH for the process being started.  The default will leave the ACS environment unchanged.
For easy access to the CASA Python interpreter, the script acsutilCASAPython
was created.  Directory paths are modified using the acsutilCASAPathInsert.py
script.  To start a container in a CASA-enabled environment, you use the
acsStartContainerWithCASA script or pass the -C option to acsutilAwaitContainerStart.
Adding the '-m casaContainer' to acsdaemonContainerStart command will trigger
the daemon to start a container with CASA support as well.

CASA modifications to the various paths set in .bash_profile.acs have been removed
in favour of the new approach.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@116363 523d945c-050c-4681-91ec-863ad3bb968a