Moreno reports the following error messages appear when the ACS environment is being setup.
sqa0 sqa-ops:~ 1001 > ssh vltuser8@alma64
Warning: No xauth data; using fake authentication data for X11 forwarding.
Last login: Thu Oct  1 07:58:30 2009 from sqa0.hq.eso.org
/introot/vltuser8/bin/python: line 133: /bin/python: No such file or directory
/introot/vltuser8/bin/python: line 133: /bin/python: No such file or directory
/introot/vltuser8/bin/python: line 133: /bin/python: No such file or directory

During the source of .bash_profile.acs, two calls to the Python interpreter
are made.  In order to these calls to use the ACS provided Python rather
than the system Python, the PATH and LD_LIBRARY_PATH variables were exported
to the shell.  The ACS Python interpreter would be picked up from the PATH
information and the commands would execute as expected.

With the introduction of the wrapper script that switches between the ACS
and CASA Python interpreters, the new script is being invoked instead of
the ACS interpreter.  The new script relies on the presence of the
PYTHON_ROOT and CASA_ROOT variables in order to determine which interpreter
to invoke.  Since PYTHON_ROOT has not been exported to the shell at that
point, the path to the interpreter generated is incorrect and the above
message will appear.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@117071 523d945c-050c-4681-91ec-863ad3bb968a