Added unit tests for the remaining parts of the Python implementation of the notification channels.  Also found an error in the ReconnectionCallback that prevented disconnection.
There are many methods in the Acspy tree that are trivial but are reported
as not being covered by the unit tests.  A number have been marked with
"# pragma: NO COVER" to exclude them from the analysis.

Renamed a number of existing unit tests so they follow the pattern used
by nose in test discovery.  The usual CVS trick of renaming the file in the
repository cannot be done in the WANDisco environment.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@124843 523d945c-050c-4681-91ec-863ad3bb968a