COMP-4212 reported that IDL files were being loaded in the wrong order, with newer files having precedence over older files.  While this is the correct order, it was discovered that there were two scripts in the system that both load the IFR.  One was written as a faster replacement for the other, but continued to exist because ITS was using it in some scripts.
This change eliminates the duplication between acsIrfeed and acsstartupLoadIFR
by removing acsIrfeed from the tree and the Makefile.  The IDL file list
generator used by acsstartupLoadIFR was listing duplicates when an IDL file
was found in multiple directories in the IDL_PATH.  The duplicates are
also eliminated from the output and some dead code was also removed.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@126853 523d945c-050c-4681-91ec-863ad3bb968a