This changes are in respect to COMP-4377: The Object Explorer was modified to use only IFR information, without needing the IDL-generated stub classes.
ObjectExplorer class was modified to remove a warning message due to Null
argument. Now an array of length 0 is passed instead.

Data encapsulation classes were implemented to handle all the complex types
without the need for the stubs. Additionally, the DataType class was implemented
to encapsulate java.lang.Class in order to store the structure definition for a
complex type (struct members, enum types, etc.).

The engine package and several classes in the BACI engine package were modified
to use this DataType class instead of the java.lang.Class in their methods.

The main modifications to allow DII calls without using the java stub classes
are located in BACIIntrospector and BACIRemoteAccess. Specifically, most changes
are encapsulated in the BACIIntrospectors methods extractAny and insertAny and
the methods used by them.

CallMethodDialog was modified to use the structure information available in a
DataType object for complex types.

The modifications in ReporterBean allow to maintain the display format used
before, with some small differences.

This changes have been tested with several types of data (primitives, sequences,
structs, enums, structSeqs, enumSeqs, etc.), but there might be some unsupported
types.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@137374 523d945c-050c-4681-91ec-863ad3bb968a