Moved the next-generation NC/DDS interfaces to here, from jcontnc (package alma.acs.nc.refactored). They still need to be improved, but this move is a first step to resolve the current build order issues.
The goal is to integrate sending and receiving of events via NC or DDS into the ACS container services. A possible future build order could be
(a) this module with the event interfaces,
(b) the module(s) with the NC/DDS implementations, that can receive ctor args ContainerServicesBase from (a) and the ORB / naming services reference.
(c) the container services implementation such as in jcont, which can then act as factory for objects from (b).


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@138706 523d945c-050c-4681-91ec-863ad3bb968a