Removed fixed castor-generated classes for the alarm schema files which were stored under com.cosylab.acs.laser.dao; using those from alma.alarmsystem.alarmmessage.generated instead, which get generated in the acsalarmidl module.
Also, separated the AlarmRefMatcher and LinkSpec classes (originally embedded into ACSAlarmDAOImpl), and refubrished them a little to keep them simple, well documented, in order to be reused (not only those, but more to come) by the cdbChecker when checking if RRs are valid in terms of alarm references and patterns.

This work is in preparation for wildcard/regexp support on the RRs alarm references. Please see the javadoc for LinkSpec for details. Tests are still passing fine, but more work needs to be done on the DAL side to continue on this


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@147743 523d945c-050c-4681-91ec-863ad3bb968a