Changes necessary to send alarms from BACI properties in the new format agreed with CONTROL. This is of the form:
 <ComponentIDLInterface#propertyName,ComponentName,code>

Currently, the ComponentIDLInterface information is taken from the corresponding "type" field on the CDB entry for the component in which the BACI property is deployed, while the propertyName corresponds to the property name as stated in the XML files under the alma/ branch of the CDB, which might be (but practically speaking shouldn't be the case because of naming conventions) the same as the IDL attribute name for the BACI property. The fact that the ComponentIDLInterface is taken from the CDB also means that, given the scenario of a hierarchy of components each defining their own BACI properties, all properties will contain the same value for ComponentIDLInterface, instead of having different ones depending on the level of the hierarchy where they were declared.

Technical corner: the ComponentIDLInterface is passed down to the BACI property through the BACIComponent, and via ContainerServices, to interfeer as less as possible with the current design of the BACI classes.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@148241 523d945c-050c-4681-91ec-863ad3bb968a