Moving Blobber Tests from module CommonSoftware/monitoring/moncontroller to CommonSoftware/monitoring/monblobber.
Some fixes were added to the tests due to changes in the interfaces and
implementation code.

History:
The files were originally copied from ARCHIVE/TMCDB/MonitorArchiver, tag
mon-refactor-201101.

There was an update from the merge done in the ARCHIVE/TMCDB/MonitorArchiver
module from several branches.

A modification for using JUnit instead of testNG was done for the tests.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@149766 523d945c-050c-4681-91ec-863ad3bb968a