New shared consumer admin usage in NCSubscriber class. Now, instead of using a global shared consumer admin object, a series of shared admin objects is used, each one in charge, theoretically, of hosting a fixed number of proxies.
To distinguish between shared consumer admin objects and those used by the old NC classes (i.e., used by only one consumer), we attach a "dummy" PUSH_ANY proxy to the shared admins when we create them.

Also, fixed the disconnect() implementation, so it disconnects the proxy supplier even if the consumer never listened to any events. After a call to disconnect(), the consumer object is unusable, and should be created again.

Added unit tests for this new feature, all of them passing, but not yet integrated with tat


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@152560 523d945c-050c-4681-91ec-863ad3bb968a