Method processEvent now gets called in a separate thread, see http://jira.alma.cl/browse/COMP-5767, so that the NC does not get blocked for a slow receiver and ACS can actually detect that a receiver is too slow. This thread is fed from a small buffer that can hold at most 8 events.
If the receiver is slow and the buffer fills up to more than half its size, then we log the LOG_NC_ReceiverTooSlow type-safe log, e.g.
   ERROR [MyComp1] More events came in from the NC than the receiver processed. [ logName='LOG_NC_ReceiverTooSlow' ClientName='MyComp1' NumEventsDiscarded='92' EventName='alma.acsnc.EventDescription' ChannelName='blarIL' Service='NotifyEventChannelFactory' LogOcurrencesNumber='100' ]
A log repeat guard is used to limit these ERROR logs to one in 30 second or for every 100 errors.

Component shutdown waits up to 2 seconds for the events in the queue to be processed.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@161176 523d945c-050c-4681-91ec-863ad3bb968a