New change on NC consumers: consumer admins are not created with names anymore, since the actual names that the TAO M&C extension uses are the proxy names. Therefore, just create the proxy with a name.
To avoid name duplication we use a random number as a suffix. Name duplication, although is anounced via exception by the M&C extension, doesn't seem to be well implemented on TAO, where the object gets first created, and the name gets checked later, with no object destruction in case of duplicated name. This could lead to memory leaks on the Notify_Service. Therefore, we try to avoid duplicated names with random suffixes for the names

This change applies to the C++ ConsumerImpl class, and both the Consumer and NCSubscriber classes on the Java world.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@166385 523d945c-050c-4681-91ec-863ad3bb968a