AlarmSourceImpl does not extends ACS::Thread anymore. The update of its internal state is updated by an external thread by calling update(...). In this way the same thread can update the state of several AlarmSourceImpl and the number of threads running in the container will be only one.
The thread for updating is passed in the constructor and the AlarmSourcImpl registers for being updated.

There is also an empty Constructor used if no external thread is available. In this case the AlarmSourceImpl object instantiates a thread. In this way the API is simpler especially when used outside of the ContainerServices


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@173023 523d945c-050c-4681-91ec-863ad3bb968a