Implemented automatic generation of a per-TMCDB-model JPA mapping file, where all TMCDB classes that map directly to a table are listed. This file can later be imported from a JPA entity-unit definition (as in ARCHIVE/TMCDB/Persistence/src/META-INF/persistence.xml) or from an hibernate configuration file (as in ARCHIVE/TMCDB/Database/src/cdb_rdb-hibernate.cfg.xml).
With this we avoid the unnecessary maintenance: a) of separate mapping files, which could be totally desynchronized, and b) of mapping files at all. Instead, we centrally list all the mappings for a given model, and then the TMCDB JPA/hibernate users just need to reference the mappings file.

I've tested this by running the hiberanteCdbJDal, and it doesn't seem to present any problem to load a configuration. I've also run the ARCHIVE/TMCDB/Persistence unit tests and they all pass. I feel confident enough, then, that the current implementation of the mappings file works reliably.


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@181161 523d945c-050c-4681-91ec-863ad3bb968a