Implemented JIRA:COMP-9338 for Java NC libs: - Added static methods to the NC Helper class: extractChannelName, extractDomainName, and combineChannelAndDomainName. - These methods are used to handle name service binding names in   - Helper#getNotificationChannel, Helper#createNotificationChannel, Helper#destroyNotificationChannel (used by supplier/subscriber etc)   - The eventGUI module, which I will commit later as part of COMP-8997. - Moved setting the NC domain name from 'Helper#getNotificationFactoryNameForChannel' to the constructor of the Helper class, as we have done already for the channel name.
The use of the NC domain name in the NameService bindings is still disabled by default, to not break tests until C++ and Py libs are changed as well.
By setting the boolean property ENFORCE_COMP_9338_IN_TRANSITION_PERIOD we can already now get the Java NC libs to use and require the new name service binding format.
TODO: Remove the handling of this property from the Java NC Helper once this ticket is implemented in all modules!


git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@188933 523d945c-050c-4681-91ec-863ad3bb968a