The login and logout from manager are executed by dedicated methods.
This change is in preparation of fixing Container.py for ICTJ:ICT-1289. BaseClient is the base class of SimpleClient and of Container. 
While SimpleClient executes disconnect to terminate, Container instead terminates with shutdown and the disconnect only logs out from the manager. So we need to decouple the logout from manger to the call of disconnect for the Container.

managerLogin(), manegerLogout() and isLoggedIn critical sections are surrounded by a Lock because then can be executed by different threads from the Container.

git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@194851 523d945c-050c-4681-91ec-863ad3bb968a