BaseClient initialised before ContainerServices.
ContainerServices should not do anything before its setAll method is invoked. But in the constructor it access the CDB (ACSCorba.cdb()) that returns the CDB with help of ACSCorba.getClient() that in turn gets the manager.

By initializing the BaseClient before ContainerServices, we ensure that the manager is accessed initially by BaseClient and we define a common path for Container and SimpleClient (see ICTJ:ICT-1289).

git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@194968 523d945c-050c-4681-91ec-863ad3bb968a