Added makeACSLogRecord, a "static" method to create a ACSLogRecord that extends logging.LogRecord.
The ACSHandler checks the timestamp of a log before sending to the logging service. If the timestamp is older then a threshold then a new record is created with a clear message and having the original message and timestampp as additional data.
The magic is done by a method called replaceOldRecord.

All the other fields of the original log remain untouched (i.e. level, source, line number and so on).
See ICTJ:ICT-548

git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@196035 523d945c-050c-4681-91ec-863ad3bb968a